requirements.txt
Pillow==8.4.0
Flask==2.0.2
celery==5.2.1
redis==4.0.2
ipython==7.29.0
flask-cors==3.0.10
